Revert "APOLLO-21277: refactored fields component"
This reverts commit 3fcba62fbec53c12ae76c2967d75f386e5cbcb89.
