+++ java-sdk/connectors/security-filtering-connector/src/main/java/com/lucidworks/fusion/connector/plugin/validation/ExampleSecurityFilterFieldsComponent.java
@@ -1,6 +1,8 @@
-package com.lucidworks.connectors.plugins.security.component;
+package com.lucidworks.fusion.connector.plugin.validation;
 
+import com.lucidworks.fusion.connector.plugin.api.security.SecurityFilter;
 import com.lucidworks.fusion.connector.plugin.api.security.SecurityFilterQueryFieldsComponent;
+import com.lucidworks.fusion.connector.plugin.api.security.Subject;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,4 +17,8 @@ public class ExampleSecurityFilterFieldsComponent implements SecurityFilterQuery
     return securityFilterFields;
   }
 
+  @Override
+  public SecurityFilter buildSecurityFilter(Subject subject) {
+    throw new UnsupportedOperationException("buildSecurityFilter is unsupported");
+  }
 }
